Dw8 xl db

+++ dw8xl/src/main/java/com/anno/dw8xl/kingdom/facade/KingdomFacade.java
@@ -6,10 +6,12 @@ package com.anno.dw8xl.kingdom.facade;
 import java.util.Collection;
 import java.util.Optional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.anno.dw8xl.dao.DAO;
+import com.anno.dw8xl.kingdom.dao.KingdomDAOInterface;
 import com.anno.dw8xl.kingdom.model.KingdomI;
 import com.anno.dw8xl.kingdom.model.NullKingdom;
 
@@ -18,39 +20,42 @@ import com.anno.dw8xl.kingdom.model.NullKingdom;
  *
  */
 @Service("kingomFacade")
-public class KingdomFacade implements KingdomFacade_I {
+public class KingdomFacade implements KingdomFacadeInterface {
 
 	@Autowired
-	private DAO<KingdomI, Object> dao;
+	private KingdomDAOInterface dao;
+	private static final Logger log = LoggerFactory.getLogger(KingdomFacade.class);
 	
 	@Override
 	public Collection<KingdomI> getAllKingdoms() {
+		log.info("Returning new ArrayList of Kingdoms...");
 		return dao.getAll();
 	}
 
 	@Override
-	public KingdomI getKingdomsByName(String name) {
-		Optional<KingdomI> res = dao.getBy(name.toUpperCase());
+	public <E> KingdomI getKingdomsBy(E criteria) {
+		Optional<KingdomI> res = dao.getBy(criteria);
 		return (res.isPresent()) ? res.get() : res.orElse(new NullKingdom());
 	}
 
 	@Override
 	public void addKingdom(KingdomI kingdom) {
+		log.info("Checking if new Kingdom already exists...");
+		if(dao.getAll().contains(kingdom)) {
+			log.info("Cannot add Kingdom due to already-existing kingdom...");
+			return;
+		}
 		dao.add(kingdom);
 	}
 
 	@Override
 	public void removeKingdom(KingdomI kingdom) {
+		log.info("Checking if kingdom exists...");
+		if(!dao.getAll().contains(kingdom)) {
+			log.info("Throwing Error due to non-existent kingdom...");
+			throw new IllegalArgumentException("Unable to Remove Kingdom: Does not Exist...");
+		}
 		dao.remove(kingdom); 
-	}
-
-//	@Override
-//	public KingdomI updateKingdom(String name, KingdomI kingdom) {
-//		List<KingdomI> temp = getKingdomsByName(name);
-//		return dao.executeUpdateKingdom(temp.get(0), kingdom);
-//		
-//	}
-
-	
+	}	
 
 }
